+++ crates/orchestrator/src/jobs/da_job/mod.rs
@@ -1,4 +1,4 @@
-use std::collections::HashMap;
+use std::collections::{HashMap, HashSet};
 use std::ops::{Add, Mul, Rem};
 use std::str::FromStr;
 use std::sync::Arc;
@@ -10,7 +10,7 @@ use lazy_static::lazy_static;
 use num_bigint::{BigUint, ToBigUint};
 use num_traits::{Num, Zero};
 use starknet::core::types::{
-    BlockId, ContractStorageDiffItem, DeclaredClassItem, Felt, MaybePendingStateUpdate, StateUpdate,
+    BlockId, ContractStorageDiffItem, DeclaredClassItem, Felt, MaybePendingStateUpdate, StateDiff, StateUpdate,
 };
 use starknet::providers::Provider;
 use thiserror::Error;
@@ -278,6 +278,7 @@ pub async fn state_update_to_blob_data(
     config: Arc<Config>,
 ) -> color_eyre::Result<Vec<Felt>> {
     let mut state_diff = state_update.state_diff;
+    refactor_state_update(&mut state_diff);
 
     let mut blob_data: Vec<Felt> = vec![Felt::from(state_diff.storage_diffs.len())];
 
@@ -392,6 +393,24 @@ fn da_word(class_flag: bool, nonce_change: Option<Felt>, num_changes: u64) -> Fe
     let decimal_string = biguint.to_str_radix(10);
 
     Felt::from_dec_str(&decimal_string).expect("issue while converting to fieldElement")
+}
+
+fn refactor_state_update(state_update: &mut StateDiff) {
+    let addresses_in_nonces: Vec<Felt> = state_update.nonces.clone().iter().map(|item| item.contract_address).collect();
+    let addresses_in_storage_diffs: Vec<Felt> =
+        state_update.storage_diffs.clone().iter().map(|item| item.address).collect();
+
+    let address_to_insert = find_unique_addresses(addresses_in_nonces, addresses_in_storage_diffs);
+
+    for address in address_to_insert {
+        state_update.storage_diffs.push(ContractStorageDiffItem { address, storage_entries: vec![] })
+    }
+}
+
+fn find_unique_addresses(nonce_addresses: Vec<Felt>, storage_diff_addresses: Vec<Felt>) -> Vec<Felt> {
+    let storage_set: HashSet<_> = storage_diff_addresses.into_iter().collect();
+
+    nonce_addresses.into_iter().filter(|addr| !storage_set.contains(addr)).collect()
 }
 
 #[cfg(test)]