+++ e2e-tests/tests.rs
@@ -46,7 +46,7 @@ struct Setup {
 
 impl Setup {
     /// Initialise a new setup
-    pub async fn new() -> Self {
+    pub async fn new(l2_block_number: String) -> Self {
         let mongo_db_instance = MongoDbServer::run().await;
         println!("✅ Mongo DB setup completed");
 
@@ -64,6 +64,7 @@ impl Setup {
         let localstack_instance = LocalStack::new().await;
         localstack_instance.setup_sqs().await.unwrap();
         localstack_instance.delete_event_bridge_rule("worker_trigger_scheduled").await.unwrap();
+        localstack_instance.setup_event_bridge(WorkerTriggerType::Snos).await.unwrap();
         localstack_instance.setup_event_bridge(WorkerTriggerType::Proving).await.unwrap();
         localstack_instance.setup_event_bridge(WorkerTriggerType::DataSubmission).await.unwrap();
         localstack_instance.setup_event_bridge(WorkerTriggerType::UpdateState).await.unwrap();
@@ -93,6 +94,7 @@ impl Setup {
             .push(("STARKNET_OPERATOR_ADDRESS".to_string(), "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266".to_string()));
         env_vec.push(("GPS_VERIFIER_CONTRACT_ADDRESS".to_string(), verifier_contract_address.to_string()));
         env_vec.push(("L1_CORE_CONTRACT_ADDRESS".to_string(), starknet_core_contract_address.to_string()));
+        env_vec.push(("MAX_BLOCK_TO_PROCESS".to_string(), l2_block_number));
 
         Self { mongo_db_instance, starknet_client, sharp_client, env_vector: env_vec, localstack_instance }
     }
@@ -127,7 +129,7 @@ async fn test_orchestrator_workflow(#[case] l2_block_number: String) {
     // setting up of the test and during orchestrator run too.
     dotenvy::from_filename(".env.test").expect("Failed to load the .env file");
 
-    let mut setup_config = Setup::new().await;
+    let mut setup_config = Setup::new(l2_block_number.clone()).await;
     // Setup S3
     setup_s3(setup_config.localstack().s3_client()).await.unwrap();